refactor(stores): break useGameplayStore <-> selectors circular import
The agent's earlier useGameplayStore decompose left a hidden circular
dep: selectors.ts imported the store hook to wrap selectors as React
hooks, while the store re-exported those hooks. Switched the design:
selectors.ts is pure functions only; hook wrappers live in the main
store file (where they can call useGameplayStore directly without
crossing the module boundary). Public API unchanged.

Tests: 600/600 pass across 19 store suites.
Tsc: clean.
